# tvgo/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a TVGO API server.",
        "title": "TVGO API Server",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/playlist": {
            "post": {
                "description": "Create playlist",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "playlist"
                ],
                "summary": "Create a new playlist",
                "operationId": "create-playlist",
                "parameters": [
                    {
                        "description": "playlist information",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CreatePlaylistDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/ID"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/playlist/page": {
            "get": {
                "description": "GetContent all playlist with pagination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "playlist"
                ],
                "summary": "GetContent all playlist with pagination",
                "operationId": "get-all-playlist",
                "parameters": [
                    {
                        "type": "string",
                        "default": "1",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "10",
                        "description": "Number of items per page",
                        "name": "perpage",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Searching by name or description",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/ID"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/playlist/search": {
            "get": {
                "description": "GetContent all playlist with pagination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "playlist"
                ],
                "summary": "GetContent all playlist with pagination",
                "operationId": "search-all-playlist",
                "parameters": [
                    {
                        "type": "boolean",
                        "name": "isVisible",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "VIP",
                            "SPORT",
                            "FREE"
                        ],
                        "type": "string",
                        "x-enum-varnames": [
                            "VIP",
                            "SPORT",
                            "FREE"
                        ],
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Limit the number of results",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/ID"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/playlist/{id}": {
            "get": {
                "description": "GetContent playlist by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "playlist"
                ],
                "summary": "GetContent playlist by ID",
                "operationId": "get-playlist-by-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "playlist ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/model.PlaylistModel"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "put playlist by ID",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "playlist"
                ],
                "summary": "put playlist",
                "operationId": "change-visibility-playlist",
                "parameters": [
                    {
                        "type": "string",
                        "description": "playlist ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete playlist by ID",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "playlist"
                ],
                "summary": "Delete playlist",
                "operationId": "delete-playlist",
                "parameters": [
                    {
                        "type": "string",
                        "description": "playlist ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Update playlist information by ID",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "playlist"
                ],
                "summary": "Update playlist information",
                "operationId": "update-playlist",
                "parameters": [
                    {
                        "type": "string",
                        "description": "playlist ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "playlist information",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UpdatePlaylistDto"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "ChannelType": {
            "type": "string",
            "enum": [
                "VIP",
                "SPORT",
                "FREE"
            ],
            "x-enum-varnames": [
                "VIP",
                "SPORT",
                "FREE"
            ]
        },
        "CreatePlaylistDto": {
            "type": "object",
            "properties": {
                "channelLink": {
                    "type": "string"
                },
                "channelType": {
                    "$ref": "#/definitions/ChannelType"
                },
                "description": {
                    "type": "string"
                },
                "logoLink": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "ID": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "UpdatePlaylistDto": {
            "type": "object",
            "properties": {
                "channelLink": {
                    "type": "string"
                },
                "channelType": {
                    "$ref": "#/definitions/ChannelType"
                },
                "description": {
                    "type": "string"
                },
                "isVisible": {
                    "type": "boolean"
                },
                "logoLink": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "model.PlaylistModel": {
            "type": "object",
            "properties": {
                "channelLink": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isVisible": {
                    "type": "boolean"
                },
                "logoLink": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}